// packages/contracts/generated/AnnotationObject.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BBox } from "./BBox";
import type { Id } from "./Id";
import type { Origin } from "./Origin";
import type { Scalar } from "./Scalar";

export type AnnotationObject = { object_id: Id, label_id: Id, geometry: BBox, attributes: { [key in string]?: Scalar }, origin: Origin, };