// packages/contracts/generated/SaveRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AnnotationDocument } from "./AnnotationDocument";
import type { Id } from "./Id";
import type { SaveLease } from "./SaveLease";
import type { SuggestionDecisionIntent } from "./SuggestionDecisionIntent";

export type SaveRequest = { operation_id: Id, base_revision_id: Id, document: AnnotationDocument, lease: SaveLease | null, suggestion_decisions: Array<SuggestionDecisionIntent>, };