// packages/contracts/generated/SaveResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AnnotationRevision } from "./AnnotationRevision";
import type { Id } from "./Id";

export type SaveResponse = { operation_id: Id, revision: AnnotationRevision, idempotent_replay: boolean, };